update (#1)
* bugfix: syscall clobbers R11

* curly parsing, proper ID types

* math parsing

* lisp parsing

* move to new file

* lisp evaluation

* working on evaluation

* printing, hooking evaluation up

* add pattern matching into the monadthis is not pleasant, but it allows us to support the "?" operation insidea match, while retaining stack traces etc.

* debugging

* set up builtins

* working on builtin implementations

* adjusting macros

* passing the basic lisp tests, add timeout and overflow

* working on elab decl; use atom ids for term/thm lookup

* typechecking expressions

* term/def elaboration

* building theorem proofs

* more readmes

* bugfixen, improved debugging

* fix comment

fixes #5

* working on refine

* finish refine and other missing functions

* bugfixing

* bugfixing, generalize printer, add check-proofs

* more debugging

* stat command

* improve synchronization

* add get-decl command, debugging

* working on add-decl

* finish implementing add-decl

* debugging lisp evaluation

* Uncons is now an OwningRef
* Added debugging printing for State and Stack
* Add variable depth on IR::Def and IR::Lambda for assertions

* more bugfixing

* prefix commands are n-ary
* uniformize dummy handling
* support inferred bound variables from deps
* parse applications in math mode

* bugfixing

* add cancellable elab, bugfix (peano.mm verifies with (check-proofs #f))

* bugfixing: peano.mm1 works!

* Added `dummy!` command for creating new dummy variables in the context
* Changed the format of `add-thm!` to take an extra argument for dummies
* Added patterns for matching on mvar and goal

* bugfix (foo (def _ bar) baz)

this should evaluate bar then (foo baz), not (foo #undef baz)

* update mm1.md and mm0-hs to MM1 changes from mm0-rs

peano.mm1 now checks with two independent verifiers

* WIP multi-file support: use async/await

* multi file bugfixing

* mm0 tweaks

* make LispVal a newtype

* use lsp versioning, use more async for scheduling

* use merkle tree hashing for file cache

* basic hover support

* hovering for formulas

* go to definition; defer proofs passed by get-decl

* document symbols command

* completion resolve request

* add hovers and go to definition for theorem applications

(now we are finally past feature parity with mm0-hs)

* add README for mm0-rs

* fix link

* oops

* pretty printed expressions

* ungroup before brackets

* adjusting grouping

* update pretty and remove patch

* standalone MM1 compiler

* [WIP] export MMB files

* example of highly redundant proof

* lean axiomatization in MM0

* demo for FOMM 2020

* bugfix lean in MM0, prove existence of nat

* more pretty errors

* add MM0 mode

* update mm0.mm0 to match arxiv paper

* bugfixes, enable diamond import

* finish export thm

* add Save command

This is needed for saving a Conv proof (which does not have a specialized
Save version).

* add mmb index

* debugging Conv instructions

* bugfix mmb output

demo.mmb passes the verifier now

* fix unfold encoding

* more annotations, DV reporting, bugfixing DVs

* more software formalization experiments

* HLA dynamic semantics

* lifting dynamic semantics to HLA

* mm0.mm0 in lean

* working on proving mm0.mm0

* procedures in refine

* simpler refine algorithm

* bugfixing

* bugfix pointer hashmap

This is a bit subtle. Although it is entirely safe rust code to have a
pointer hashmap, the keys may not be pointing to valid memory,
and in particular if the memory is reused, a different object might
compare equal to the original, dropped value.

To fix this, we let the hashmap own the keys so that they will not move
or be dropped until the pretty printer is done. (Alternatively, we
could have a fancy key type that is (*const T, Arc<T>) but this
conflicts with our existing newtype for LispVal.)

* finished section 1 of mm0.mm1

* fix mm0 grammar

* add "mm0-rs join" utility to remove "import" lines from mm0

* fix dummies, add theorem dummy check

* add compilation instructions to README

fixes #9

* restore correct printer for mm0-hs from-mm

fixes #10

* tweak wording (fixes #11)

* local variables shadow term constructors, even if non-nullary

* add mmu exporter with demo output

* add mmu import

Now you can `import "foo.mmu"` inside an mm1 file, and you can also
use an mmu file as the source for a compile command. In particular,
this enables converting mmu proofs into mmb format
using `mm0-rs compile foo.mmu foo.mmb`.

* bugfix x86 spec: condition codes

* bugfixes, add dummy clause to :unfold

* bring peano.mm0 in line with peano.mm1

as observed by @ammkrn, we finally have a way to compare mm1 files to
mm0 as originally intended, by using mm0-rs to compile mm1 -> mmu
and using mm0-hs to verify mmu against mm0. As expected, there are a
bunch of small issues to fix.

Some of these changes are apparently superfluous, because the mm0-hs
verifier does not support matching up to alpha renaming. I haven't decided whether
to relax the verifier, but it's not hard to just change names to get
exact equality here.

* bugfix: don't check dv conditions in unfold params

This is actually a bug in the paper version of MM0, reported by
Asger Ipsen. If you enforce dv conditions between parameters, then terms
like `[x + 1 / x] ph` become stuck (not unfoldable) because the
unfolding step would enforce that x+1 and x are disjoint.

* fix proofs

the MM1 dv checker is a bit inaccurate, and let these errors go through.
Independent proof checking FTW!

* bugfix x86 model: 64 bit immediates act as 32 bit

* Update vscode-mm0/README.md

* bugfix x86 spec

* fix precedences

Because it is an error for a precedence level to be both left- and
right-associative, we use the convention that left assoc precedences
have even numbers, and right assoc have odd numbers.

* add MM0 parsing to mm0-c

This is long overdue, an essential part of mm0-c for correctness, but also a
fairly hefty part of the code. The old behavior of the verifier is still
accessible via the NO_PARSER preprocessor directive, in which case it will
ignore standard in and only check the proof for internal consistency.

* update mm0-c readme

* mm0-hs support for generalized infix notations

Co-authored-by: Mario Carneiro <di.gama@gmail.com>
Co-authored-by: Johan Commelin <johan@commelin.net>